+++ lib/Eclair/AST/Analysis.hs
@@ -9,7 +9,6 @@ module Eclair.AST.Analysis
   , runAnalysis
   , VariableInFact(..)
   , UngroundedVar(..)
-  , EmptyModule(..)
   , WildcardInFact(..)
   , WildcardInRuleHead(..)
   , WildcardInAssignment(..)
@@ -194,12 +193,6 @@ data WildcardInAssignment
   deriving anyclass S.Marshal
   deriving S.Fact via S.FactOptions WildcardInAssignment "wildcard_in_assignment" 'S.Output
 
-newtype EmptyModule
-  = EmptyModule NodeId
-  deriving stock (Generic, Eq, Show)
-  deriving anyclass S.Marshal
-  deriving S.Fact via S.FactOptions EmptyModule "empty_module" 'S.Output
-
 newtype DeadCode
   = DeadCode { unDeadCode :: NodeId }
   deriving stock (Generic, Eq, Show)
@@ -242,7 +235,6 @@ data SemanticAnalysis
        , PointsToVar
        , VariableInFact
        , UngroundedVar
-       , EmptyModule
        , WildcardInRuleHead
        , WildcardInFact
        , WildcardInAssignment
@@ -282,8 +274,7 @@ data Result
 
 data SemanticErrors
   = SemanticErrors
-  { emptyModules :: Container EmptyModule
-  , variablesInFacts :: Container VariableInFact
+  { variablesInFacts :: Container VariableInFact
   , ungroundedVars :: Container UngroundedVar
   , wildcardsInFacts :: Container WildcardInFact
   , wildcardsInRuleHeads :: Container WildcardInRuleHead
@@ -295,7 +286,6 @@ data SemanticErrors
 
 hasSemanticErrors :: Result -> Bool
 hasSemanticErrors result =
-  isNotNull emptyModules ||
   isNotNull variablesInFacts ||
   isNotNull ungroundedVars ||
   isNotNull wildcardsInFacts ||
@@ -374,7 +364,6 @@ analysis prog = S.mkAnalysis addFacts run getFacts
       info <- SemanticInfo <$> (mkPointsToAnalysis <$> S.getFacts prog)
                            <*> S.getFacts prog
       errs <- SemanticErrors <$> S.getFacts prog
-                             <*> S.getFacts prog
                              <*> S.getFacts prog
                              <*> S.getFacts prog
                              <*> S.getFacts prog